Prefer sources labeled "Pragmatux" over others
The idea is to override packages in Debian which also have versions in
the Pragmatux sources. We sometimes put a modified version of a Debian
package in Pragmatux in order to fix a bug or add a feature. When
making such a modification, we increment the package version with a
suffix like "ptux1" (see Ubuntu's scheme for doing the same).  This
temporarily causes our package to win over Debian's; however, if
Debian were to release a new version, their package would win until we
rereleased our package with a newer version number. This
priority-pinning preference causes the Pragmatux version to win
regardless of the version number in Debian.

Our practice is to send bug fixes and feature enhancements upstream to
Debian such that our changes eventually end up in Debian's version of
the package. Once they do, we remove the package from the Pragmatux
repositories, unmasking Debian's version of the package.